Show upload progress bar in shell_example
